Prepare v2015.10

- Fix defconfig
- Fix qspi support for cc108
- Enable NET_RANDOM_ETHADDR for mainline boards
- Fix USB GADGET and LTHOR macros

+++ arch/arm/mach-tegra/pinmux-common.c
@@ -78,7 +78,7 @@
 	(((hsm) >= PMUX_HSM_DISABLE) && ((hsm) <= PMUX_HSM_ENABLE))
 #endif
 
-#define _R(offset)	(u32 *)(NV_PA_APB_MISC_BASE + (offset))
+#define _R(offset)	(u32 *)((unsigned long)NV_PA_APB_MISC_BASE + (offset))
 
 #if defined(CONFIG_TEGRA20)
